更新脚本，更新更多接口，优化命名缺陷

+++ runtime/scripts/utils/Item/BaseWeapon.cs
@@ -3,29 +3,29 @@ using Godot;
 [GlobalClass]
 public partial class BaseWeapon : BaseItem
 {
-    [Export]
-    public float Attack = 0.0F;
-    [Export]
-    public PackedScene AttackAnimation;
+	[Export]
+	public float Attack = 0.0F;
+	[Export]
+	public PackedScene AttackAnimation;
 
-    public virtual float _CalculateDamage(float value, BaseEnemy targetEnemy)
-    {
-        float atk = UtmxPlayerDataManager.PlayerAttack + Attack;
-        float def = targetEnemy.Defence;
-        float damage = atk - def + (GD.Randf() * 2);
-        if (value <= 12)
-        {
-            damage *= 2.2F;
-        }
-        else
-        {
-            damage *= (1 - value / 545) * 2F;
-        }
-        damage = Mathf.Round(damage);
-        if (damage <= 0)
-        {
-            damage = 0;
-        }
-        return damage;
-    }
+	public virtual double _CalculateDamage(float value, BaseEnemy targetEnemy)
+	{
+		double atk = UtmxPlayerDataManager.PlayerAttack + Attack;
+		double def = targetEnemy.Defence;
+		double damage = atk - def + (GD.Randf() * 2);
+		if (value <= 12)
+		{
+			damage *= 2.2F;
+		}
+		else
+		{
+			damage *= (1 - value / 545) * 2F;
+		}
+		damage = Mathf.Round(damage);
+		if (damage <= 0)
+		{
+			damage = 0;
+		}
+		return damage;
+	}
 }
